ref: check_envelope
